requirements.txt
requests>=2.30.0
python-dotenv>=1.0.0
openai>=1.3.0
tqdm>=4.65.0
numpy>=1.24.0
scikit-learn>=1.2.0
python-dateutil>=2.8.0
tiktoken>=0.5.0
